Move stream pytest fixtures into conftest.py
